+++ packages/extended-text-entry/configure/src/main.jsx
@@ -67,15 +67,15 @@ export class Main extends React.Component {
   render() {
     const { model, classes, onModelChanged, configuration, onConfigurationChanged, imageSupport } = this.props;
     const {
-      multiple,
-      feedback,
+      multiple = {},
+      feedback = {},
       teacherInstructions = {},
       prompt = {},
-      studentInstructions,
-      mathInput,
-      dimensions,
-      equationEditor
-    } = configuration;
+      studentInstructions = {},
+      mathInput = {},
+      dimensions = {},
+      equationEditor = {}
+    } = configuration || {};
 
     return (
       <layout.ConfigLayout
@@ -129,10 +129,7 @@ export class Main extends React.Component {
         }
       >
         <div>
-          <Typography className={classes.header} variant="subheading">
-            Display
-          </Typography>
-
+          
           {teacherInstructions.enabled && (
             <InputContainer label={teacherInstructions.label} className={classes.promptContainer}>
               <EditableHtml