+++ packages/math-inline/configure/src/configure.jsx
@@ -67,13 +67,13 @@ export class Configure extends React.Component {
     const { classes, model, imageSupport, onModelChanged, configuration, onConfigurationChanged } = this.props;
     const {
       feedback = {},
-      responseType,
+      responseType = {},
       teacherInstructions = {},
-      studentInstructions,
-      rationale,
+      studentInstructions = {},
+      rationale = {},
       prompt = {},
-      scoringType
-    } = configuration;
+      scoringType = {}
+    } = configuration || {};
     log('[render] model', model);
 
 
@@ -111,13 +111,6 @@ export class Configure extends React.Component {
         >
           <div>
             <div className={classes.content}>
-              <Typography component="div" type="body1">
-            <span>
-              In Inline Math, students need to fill in the blank for an equation or a mathematical expression.
-              This interaction allows for exactly one correct answer.
-            </span>
-              </Typography>
-
               {teacherInstructions.enabled && (
                 <InputContainer label={teacherInstructions.label} className={classes.promptHolder}>
                   <EditableHtml