+++ packages/charting/CHANGELOG.md
@@ -2,3 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# 0.1.0 (2019-08-15)
+
+
+### Features
+
+* Added pie-model. ([30e7c46](https://github.com/pie-framework/pie-elements/commit/30e7c46))
+* Send corrected answer when 'evaluate' or 'view' mode is enabled. ([d540568](https://github.com/pie-framework/pie-elements/commit/d540568))
+* **charting:** Added charting package and started implementation (render charts, select chart type, define correct answer). (WIP) ([75efd72](https://github.com/pie-framework/pie-elements/commit/75efd72))
+* **charting:** Scoring. ([b3b97e3](https://github.com/pie-framework/pie-elements/commit/b3b97e3))
+* **charting:** Scoring. ([6088c42](https://github.com/pie-framework/pie-elements/commit/6088c42))
+* **charting:** Updated Scoring. ([e2ccf04](https://github.com/pie-framework/pie-elements/commit/e2ccf04))