+++ packages/graphing/controller/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/graphing-controller@1.2.4...@pie-element/graphing-controller@1.2.5) (2019-08-15)
+
+**Note:** Version bump only for package @pie-element/graphing-controller
+
+
+
+
+
+## [1.2.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/graphing-controller@1.2.3...@pie-element/graphing-controller@1.2.4) (2019-08-02)
+
+**Note:** Version bump only for package @pie-element/graphing-controller
+
+
+
+
+
+## [1.2.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/graphing-controller@1.2.2...@pie-element/graphing-controller@1.2.3) (2019-07-25)
+
+**Note:** Version bump only for package @pie-element/graphing-controller
+
+
+
+
+
 ## [1.2.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/graphing-controller@1.2.1...@pie-element/graphing-controller@1.2.2) (2019-07-07)
 
 **Note:** Version bump only for package @pie-element/graphing-controller