+++ packages/image-cloze-association/controller/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-controller@1.2.0...@pie-element/image-cloze-association-controller@1.2.1) (2019-08-15)
+
+**Note:** Version bump only for package @pie-element/image-cloze-association-controller
+
+
+
+
+
+# [1.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-controller@1.1.2...@pie-element/image-cloze-association-controller@1.2.0) (2019-08-02)
+
+
+### Bug Fixes
+
+* calculate correct partial when comes to duplicates and extra answers ([a8b1677](https://github.com/pie-framework/pie-elements/commit/a8b1677))
+
+
+### Features
+
+* support alternate correct answers ([03b5ce0](https://github.com/pie-framework/pie-elements/commit/03b5ce0))
+
+
+
+
+
+## [1.1.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-controller@1.1.1...@pie-element/image-cloze-association-controller@1.1.2) (2019-07-25)
+
+**Note:** Version bump only for package @pie-element/image-cloze-association-controller
+
+
+
+
+
 ## [1.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-controller@1.1.0...@pie-element/image-cloze-association-controller@1.1.1) (2019-07-11)
 
 **Note:** Version bump only for package @pie-element/image-cloze-association-controller