+++ packages/match-list/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-list@1.5.4...@pie-element/match-list@1.5.5) (2019-08-15)
+
+**Note:** Version bump only for package @pie-element/match-list
+
+
+
+
+
+## [1.5.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-list@1.5.3...@pie-element/match-list@1.5.4) (2019-08-02)
+
+
+### Bug Fixes
+
+* **elements:** fixed some feedback config issues and the prompt issue on categorize [ch2192], [ch1770] ([eb22724](https://github.com/pie-framework/pie-elements/commit/eb22724))
+
+
+
+
+
+## [1.5.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-list@1.5.2...@pie-element/match-list@1.5.3) (2019-07-25)
+
+**Note:** Version bump only for package @pie-element/match-list
+
+
+
+
+
 ## [1.5.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-list@1.5.1...@pie-element/match-list@1.5.2) (2019-07-07)
 
 **Note:** Version bump only for package @pie-element/match-list