+++ packages/match/controller/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.9](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@2.2.8...@pie-element/match-controller@2.2.9) (2019-08-15)
+
+**Note:** Version bump only for package @pie-element/match-controller
+
+
+
+
+
+## [2.2.8](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@2.2.7...@pie-element/match-controller@2.2.8) (2019-08-02)
+
+
+### Bug Fixes
+
+* **elements:** fixed some feedback config issues and the prompt issue on categorize [ch2192], [ch1770] ([eb22724](https://github.com/pie-framework/pie-elements/commit/eb22724))
+
+
+
+
+
+## [2.2.7](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@2.2.6...@pie-element/match-controller@2.2.7) (2019-07-19)
+
+**Note:** Version bump only for package @pie-element/match-controller
+
+
+
+
+
 ## [2.2.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@2.2.5...@pie-element/match-controller@2.2.6) (2019-07-17)