+++ packages/multiple-choice/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.8](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@3.5.7...@pie-element/multiple-choice@3.5.8) (2019-08-15)
+
+
+### Bug Fixes
+
+* **elements:** version bumped the editable-html packages [ch1738] ([949222a](https://github.com/pie-framework/pie-elements/commit/949222a))
+
+
+
+
+
+## [3.5.7](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@3.5.6...@pie-element/multiple-choice@3.5.7) (2019-08-02)
+
+
+### Bug Fixes
+
+* **elements:** fixed some feedback config issues and the prompt issue on categorize [ch2192], [ch1770] ([eb22724](https://github.com/pie-framework/pie-elements/commit/eb22724))
+
+
+
+
+
+## [3.5.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@3.5.5...@pie-element/multiple-choice@3.5.6) (2019-07-25)
+
+**Note:** Version bump only for package @pie-element/multiple-choice
+
+
+
+
+
+## [3.5.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@3.5.4...@pie-element/multiple-choice@3.5.5) (2019-07-25)
+
+**Note:** Version bump only for package @pie-element/multiple-choice
+
+
+
+
+
 ## [3.5.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@3.5.3...@pie-element/multiple-choice@3.5.4) (2019-07-13)
 
 **Note:** Version bump only for package @pie-element/multiple-choice