+++ packages/point-intercept/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/point-intercept@4.4.5...@pie-element/point-intercept@4.4.6) (2019-08-15)
+
+
+### Bug Fixes
+
+* **elements:** version bumped the editable-html packages [ch1738] ([949222a](https://github.com/pie-framework/pie-elements/commit/949222a))
+
+
+
+
+
+## [4.4.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/point-intercept@4.4.4...@pie-element/point-intercept@4.4.5) (2019-08-02)
+
+**Note:** Version bump only for package @pie-element/point-intercept
+
+
+
+
+
+## [4.4.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/point-intercept@4.4.3...@pie-element/point-intercept@4.4.4) (2019-07-25)
+
+**Note:** Version bump only for package @pie-element/point-intercept
+
+
+
+
+
 ## [4.4.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/point-intercept@4.4.2...@pie-element/point-intercept@4.4.3) (2019-07-12)
 
 **Note:** Version bump only for package @pie-element/point-intercept